Warning set to 3 for windows builds Disabled warning 4251 on windows as it is given by exporting classes or structures with stl member variables (this is perfectly safe for our use case) emoji_flags set to uint8_t - binary operators used on signed type role_flags set to uint8_t - binary operators used on signed type cluster constructor exception formal parameter removed as its unused lots of type conversions forward declarations removed from commandhandler.cpp changing NULL to nullptr removed ifs around deletes as deleting nullptr does nothing making the check unnecessary changing c headers to c++ versions of c headers (errno.h -> cerrno ect) removed struct before type from socket code as this is legacy c code and no longer needed utility.cpp removed unnecessary return added "u" to end of numeric values to define as unsigned Added warning disable to test for same reasons as added to dpp

+++ src/dpp/wsclient.cpp
@@ -26,8 +26,8 @@
 
 namespace dpp {
 
-const unsigned char WS_MASKBIT = (1 << 7);
-const unsigned char WS_FINBIT = (1 << 7);
+const unsigned char WS_MASKBIT = (1u << 7u);
+const unsigned char WS_FINBIT = (1u << 7u);
 const unsigned char WS_PAYLOAD_LENGTH_MAGIC_LARGE = 126;
 const unsigned char WS_PAYLOAD_LENGTH_MAGIC_HUGE = 127;
 const size_t WS_MAX_PAYLOAD_LENGTH_SMALL = 125;
@@ -80,7 +80,7 @@ size_t websocket_client::FillHeader(unsigned char* outbuf, size_t sendlength, ws
 
 	if (sendlength <= WS_MAX_PAYLOAD_LENGTH_SMALL)
 	{
-		outbuf[pos++] = sendlength;
+		outbuf[pos++] = (unsigned int)sendlength;
 	}
 	else if (sendlength <= WS_MAX_PAYLOAD_LENGTH_LARGE)
 	{
